Centralize Screen_TickFlash prototype into include/gfx.h
Binary diagnosis: all 9 baserom call sites show bare cmp r0,#0 with
no lsls #24 zero-extend — the true return type is s32, not u8. Changed
definition return type u8->s32 (byte-neutral: body returns *countdown
via ldrb which already zero-extends; no exit-mask exists). 7 callers
de-externed; also migrated sub_08017bfc.c (int Screen_TickFlash;
without extern keyword) and sub_08017abc.c to use the header.

+++ src/engine/sub_080106b8.c
@@ -58,7 +58,7 @@ u16 Screen_BeginFlash(u16 arg)
  * final block doesn't need the bldcnt/n pins because there's no
  * arithmetic on countdown before the store. */
 
-u8 Screen_TickFlash(void)
+s32 Screen_TickFlash(void)
 {
     vu8 *countdown = (vu8 *)0x03006538;
     vu8 *p3480 = (vu8 *)0x03003480;
